Adding another project to wrap ASIHTTPRequest into it's own project with products for Mac OS X and iOS.
This is because it's likely that BSCouchObjC will be used in other projects. In and of itself it isn't at all useful. So, building the ASIHTTPRequest classes into the framework/library isn't very flexible, as more than likely the main project will want to use ASIHTTPReqest too. So, instead we can package the framework up separately, but don't link against it on iOS, and don't copy it on Mac OS X.

Signed-off-by: Daniel Thorpe <dan@blindingskies.com>